refactor: relocate the player-related functions
