+++ langchain_werewolf/game_players/__init__.py
@@ -7,6 +7,9 @@ from .const import (
 )
 from .registry import PlayerSideRegistry, PlayerRoleRegistry
 from .utils import (
+    find_player_by_name,
+    find_players_by_role,
+    find_players_by_side,
     is_player_with_side,
     is_player_with_role,
     is_valid_game_player,
@@ -20,6 +23,9 @@ __all__ = [
     BaseGamePlayerRole.__name__,
     PlayerSideRegistry.__name__,
     PlayerRoleRegistry.__name__,
+    find_player_by_name.__name__,
+    find_players_by_role.__name__,
+    find_players_by_side.__name__,
     is_player_with_side.__name__,
     is_player_with_role.__name__,
     is_valid_game_player.__name__,
