+++ tests/test_utils.py
@@ -1,18 +1,8 @@
 from typing import Callable, TypeVar
 import pytest
 from pytest_mock import MockerFixture
-from langchain_core.runnables import RunnableLambda
-from langchain_werewolf.game_players import BaseGamePlayer
-from langchain_werewolf.game_players.player_roles import (
-    Villager,
-    Werewolf,
-)
-from langchain_werewolf.game_players.registry import PlayerRoleRegistry
 from langchain_werewolf.utils import (
     assert_not_empty_deco,
-    find_player_by_name,
-    find_players_by_role,
-    find_players_by_side,
     consecutive_string_generator,
     random_permutated_infinite_generator,
     remove_none_values,
@@ -29,232 +19,6 @@ def test_assert_not_empty_deco():
     assert test_func([1, 2, 3]) == [1, 2, 3]
     with pytest.raises(ValueError):
         test_func([])
-
-
-@pytest.mark.parametrize(
-    'name, players, expected',
-    [
-        (
-            'Alice',
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            PlayerRoleRegistry.create_player(
-                key=Villager.role,
-                name='Alice',
-                runnable=RunnableLambda(str),
-            ),
-        ),
-        (
-            'Bob',
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            PlayerRoleRegistry.create_player(
-                key=Werewolf.role,
-                name='Bob',
-                runnable=RunnableLambda(str),
-            ),
-        ),
-    ],
-)
-def test_find_player_by_name(
-    name: str,
-    players: list[BaseGamePlayer],
-    expected: BaseGamePlayer,
-):
-    assert find_player_by_name(name, players) == expected
-
-
-def test_find_player_by_name_not_found():
-    with pytest.raises(ValueError):
-        find_player_by_name('Alice', [])
-
-
-def test_find_player_by_name_not_unique():
-    with pytest.raises(ValueError):
-        find_player_by_name(
-            'Alice',
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-            ] * 3,
-        )
-
-
-@pytest.mark.parametrize(
-    'role, players, expected',
-    [
-        (
-            Villager.role,
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-        ),
-        (
-            Werewolf.role,
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-        ),
-    ],
-)
-def test_find_players_by_role(
-    role: str,
-    players: list[BaseGamePlayer],
-    expected: list[BaseGamePlayer],
-):
-    assert find_players_by_role(role, players) == expected
-
-
-def test_find_players_by_role_not_found():
-    with pytest.raises(ValueError):
-        find_players_by_role(Villager.role, [])
-
-
-@pytest.mark.parametrize(
-    'side, players, expected',
-    [
-        (
-            Villager.side,
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-        ),
-        (
-            Werewolf.side,
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Alice',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-                PlayerRoleRegistry.create_player(
-                    key=Villager.role,
-                    name='Charlie',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-            [
-                PlayerRoleRegistry.create_player(
-                    key=Werewolf.role,
-                    name='Bob',
-                    runnable=RunnableLambda(str),
-                ),
-            ],
-        ),
-    ],
-)
-def test_find_players_by_side(
-    side: str,
-    players: list[BaseGamePlayer],
-    expected: list[BaseGamePlayer],
-):
-    assert find_players_by_side(side, players) == expected
-
-
-def test_find_players_by_side_not_found():
-    with pytest.raises(ValueError):
-        find_players_by_side(Villager.side, [])
 
 
 @pytest.mark.parametrize(
